# docs/swagger.yaml
basePath: /
definitions:
...
  models.Achievement:
    properties:
      approved:
        type: boolean
      createdAt:
        type: string
      createdBy:
        type: integer
      description:
        type: string
      id:
        type: string
      points:
        type: number
      title:
        type: string
    type: object
...
host: 109.73.202.151:8080
info:
  contact: {}
  description: ITaM API
  title: LiveCode API
  version: "1.0"
paths:
  /auth/api/create_achievement:
    post:
      consumes:
      - application/json
      description: Создает новое достижение
      parameters:
      - description: Achievement data
        in: body
        name: achievement
        required: true
        schema:
          $ref: '#/definitions/models.Achievement'
      produces:
      - application/json
      responses:
        "201":
          description: Success message with ID
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid title or points
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to save achievement
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Создать достижение
      tags:
      - Achievements
...
  /auth/api/delete_achievement:
    delete:
      description: Удаляет достижение по его ID
      parameters:
      - description: Achievement ID
        in: query
        name: achievement_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid achievement ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Удалить достижение
      tags:
      - Achievements
...
  /auth/api/get_achievement:
    get:
      description: Возвращает информацию о конкретном достижении
      parameters:
      - description: Achievement ID
        in: query
        name: achievement_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Achievement data
          schema:
            $ref: '#/definitions/models.Achievement'
        "400":
          description: Invalid achievement ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Achievement not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error while fetching achievement
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить достижение
      tags:
      - Achievements
  /auth/api/get_all_achievements:
    get:
      description: Возвращает список всех достижений с пагинацией
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of all achievements
          schema:
            items:
              $ref: '#/definitions/models.Achievement'
            type: array
        "400":
          description: Invalid pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить все достижения
      tags:
      - Achievements
...
  /auth/api/get_user_achievements:
    get:
      description: Возвращает список достижений пользователя с пагинацией
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of achievements
          schema:
            items:
              $ref: '#/definitions/models.Achievement'
            type: array
        "400":
          description: Invalid user ID or pagination parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Получить достижения пользователя
      tags:
      - Achievements
...
  /auth/api/update_achievement:
    patch:
      consumes:
      - application/json
      description: Обновляет существующее достижение
      parameters:
      - description: Achievement data
        in: body
        name: achievement
        required: true
        schema:
          $ref: '#/definitions/models.Achievement'
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid achievement ID or data
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - OAuth2Password: []
      summary: Обновить достижение
      tags:
      - Achievements
...
schemes:
- http
securityDefinitions:
  OAuth2Password:
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"